host_logic -> host_api
